requirements.txt
Flask~=3.0.2
pytest~=8.0.0
PyJWT~=2.8.0
requests~=2.31.0
python-dotenv~=1.0.1